Refactor `ChangesetOverwriteStrategy` to inherit an abstract class
We need a second `ChangesetApplicationStrategy`, so this extracts the general
interface to an abstract class and tests it in shared examples.
