+++ lib/tufts/changeset_preserve_strategy.rb
@@ -1,0 +1,2 @@
+module Tufts
+end
