+++ spec/lib/tufts/changeset_application_strategy_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Tufts::ChangesetApplicationStrategy do
+  it_behaves_like 'a ChangesetApplicationStrategy'
+end
