+++ spec/lib/tufts/changeset_overwrite_strategy_spec.rb
@@ -4,30 +4,11 @@ RSpec.describe Tufts::ChangesetOverwriteStrategy do
   subject(:strategy) { described_class.new(model: model) }
   let(:model)        { FakeWork.new }
 
-  it { is_expected.to have_attributes(model: model, changeset: be_empty) }
-
-  shared_context 'with changes' do
-    subject(:strategy) do
-      described_class.new(model: model, changeset: changeset)
-    end
-
-    let(:changeset) do
-      new_model = FakeWork.new(title: ['moomin', 'moominmama', 'snork'],
-                               subject: ['too-ticky'])
-      ActiveFedora::ChangeSet
-        .new(new_model, new_model.resource, new_model.changed_attributes.keys)
-    end
-  end
+  it_behaves_like 'a ChangesetApplicationStrategy'
 
   describe '#apply' do
-    context 'with no changes' do
-      it 'leaves the model unchanged' do
-        expect { strategy.apply }.not_to change { model.resource.statements }
-      end
-    end
-
     context 'with changes' do
-      include_context 'with changes'
+      include_context 'strategy with changes'
 
       it 'applies the changes' do
         expect { strategy.apply }
